version 2

+++ controllers/likeController.js
@@ -1,3 +1,4 @@
+const Post = require("../models/Post");
 const Like = require("../models/Like");
 const Dislike = require("../models/Dislike");
 
@@ -5,30 +6,36 @@ exports.likePost = async (req, res) => {
   const postId = req.params.postId;
   const userId = req.user._id;
 
+  if(!userId) {
+    return res.status(401).json({ status: "error", message: "User not logged in" });
+  }
+
   try {
-    // Check if user liked the post
+    const post = await Post.findById(postId);
+
     const existingLike = await Like.findOne({ post: postId, user: userId });
 
     if (existingLike) {
-      // If liked remove the like
       await Like.findOneAndDelete({ post: postId, user: userId });
+      post.likeCount -= 1;
+      await post.save();
       return res
         .status(200)
         .json({ status: "success", message: "Like removed successfully!" });
     }
 
-    // Remove the dislike
     await Dislike.findOneAndDelete({ post: postId, user: userId });
 
-    // Like
     const newLike = new Like({ post: postId, user: userId });
     await newLike.save();
+    post.likeCount += 1;
+    await post.save();
 
-    return res
+    res
       .status(201)
       .json({ status: "success", message: "Post liked successfully." });
   } catch (error) {
-    return res.status(500).json({ status: "error", message: error.message });
+    res.status(500).json({ status: "error", message: error.message });
   }
 };
 
@@ -37,6 +44,8 @@ exports.dislikePost = async (req, res) => {
   const userId = req.user._id;
 
   try {
+    const post = await Post.findById(postId);
+
     const existingDislike = await Dislike.findOne({
       post: postId,
       user: userId,
@@ -44,6 +53,8 @@ exports.dislikePost = async (req, res) => {
 
     if (existingDislike) {
       await Dislike.findOneAndDelete({ post: postId, user: userId });
+      post.dislikeCount -= 1;
+      await post.save();
       return res
         .status(200)
         .json({ status: "success", message: "Dislike removed successfully." });
@@ -53,9 +64,11 @@ exports.dislikePost = async (req, res) => {
 
     const newDislike = new Dislike({ post: postId, user: userId });
     await newDislike.save();
+    post.dislikeCount += 1;
+    await post.save();
 
-    return res.status(201).json({ message: "Post disliked successfully." });
+    res.status(201).json({ message: "Post disliked successfully." });
   } catch (error) {
-    return res.status(500).json({ status: "error", message: error.message });
+    res.status(500).json({ status: "error", message: error.message });
   }
 };
